pygments styles added few fixes

+++ _posts/2013-08-26-raspberry-pi-ssh-destegi-ile-boot.markdown
@@ -27,57 +27,51 @@ Bilindiği gibi Raspberry Pi için oluşturulan distrolar monitör ve klavye ile
 Linux makinelere ssh protokolü ile bağlanıp bash'te komut çalıştırılabilir. Ancak makine'de ssh server bulunmalı ve bu makinede bir takım konfigürasyonlar yapılmalıdır. Bu konfigurasyonlar monitör olmadığı için yükleme yapılmış sd kartta boot öncesi yapılmak zorunda. Bir takım araştırmalar sonucu elde ettiğim bilgileri adım adım yapılacaklar biçiminde derledim.<!-- more -->
 
 
-Raspbian wheezy imajı (Pi için Debian distrosu) kurulu bir sdkartın hazır olduğunu varsayarak, _Ubuntu/Linux üzerinde_, sonraki adımlara göz atabiliriz. (Raspbian kurulumu için [tıklayınız](http://elinux.org/RPi_Easy_SD_Card_Setup#Using_command_line_tools_.281.29).)
+Raspbian wheezy imajı (Pi için Debian distrosu) kurulu bir sdkartın hazır olduğunu varsayarak, *Ubuntu/Linux üzerinde* sonraki adımlara göz atabiliriz. (Raspbian kurulumu için [tıklayınız](http://elinux.org/RPi_Easy_SD_Card_Setup#Using_command_line_tools_.281.29).)
 
 
 Raspbian kurulu sdkart üzerinde iki partition bulunur. Bunlardan birisinde bin, dev, etc.. gibi dizinler var. Bu partition üzerinde işlem yapılacaktır. Verilen dizin yapısı buna göre değerlendirilmelidir.
 
 Pi networke bağlandığında, statik IP alması için /etc/network/interfaces dosyasında değişiklik yapılmalıdır.
 
-    
-    sudo cp etc/network/interfaces etc/network/interfaces.bak
-    # dosya yedeklenir
-    sudo gedit etc/network/interfaces
-    # metin editörü ile interfaces dosyası açılır.
-
+{% highlight bash %}
+sudo cp etc/network/interfaces etc/network/interfaces.bak # dosya yedeklenir
+sudo gedit etc/network/interfaces # metin editörü ile interfaces dosyası açılır.
+{% endhighlight %}
 
 Dosyanın içeriği aşağıdaki içerik ile değiştirilir. Network gereksinimlerine göre IP adreslerinde değişiklikler yapılmalıdır.
 
-    
-    auto lo eth0
-    iface default inet dhcp
-    iface lo inet loopback
-    iface eth0 inet static
-    address 192.168.1.67 # istenen statik ip
-    netmask 255.255.255.0
-    gateway 192.168.1.1
-
+{% highlight bash %}
+auto lo eth0
+iface default inet dhcp
+iface lo inet loopback
+iface eth0 inet static
+address 192.168.1.67 # istenen statik ip
+netmask 255.255.255.0
+gateway 192.168.1.1
+{% endhighlight %}
 
 Artık Pi boot edildiğinde verilen statik ip ile networke bağlanacaktır.
 
-SSH Raspbian içerisinde aktifleştirilmiş olarak bulunuyor ancak bir takım konfigürasyonlar yapmak gerekmekte. Bunun için bash'te çalışacak komutlar yazmak mümkün. Bu çözüme [stackexhange](http://raspberrypi.stackexchange.com/questions/4444/enabling-ssh-on-rpi-without-screen-keystrokes-for-raspi-config#answer-8083)'de [nortally](http://raspberrypi.stackexchange.com/users/8114/nortally) tarafından yazılımış bir cevap ile ulaşmıştım.]
+SSH Raspbian içerisinde aktifleştirilmiş olarak bulunuyor ancak bir takım konfigürasyonlar yapmak gerekmekte. Bunun için bash'te çalışacak komutlar yazmak mümkün. Bu çözüme [stackexhange](http://raspberrypi.stackexchange.com/questions/4444/enabling-ssh-on-rpi-without-screen-keystrokes-for-raspi-config#answer-8083)'de [nortally](http://raspberrypi.stackexchange.com/users/8114/nortally) tarafından yazılımış bir cevap ile ulaşmıştım.
 
 /etc dizininde bayrak niteliğinde bir dosya eklenir:
 
-    
-    sudo touch etc/SSHFLAG
-
-
-
-    
-    sudo gedit etc/rc.local       # metin editöründe dosya açılır.
+{% highlight bash %}
+sudo touch etc/SSHFLAG
+sudo gedit etc/rc.local       # metin editöründe dosya açılır.
+{% endhighlight %}
 
 
 Editörde:
 
-    
-    <code>if [ -e /etc/SSHFLAG ]; then
-      /usr/sbin/update-rc.d -f ssh defaults
-      /bin/rm /etc/SSHFLAG
-      /sbin/shutdown -r now
-    fi
-    </code>
-
+{% highlight bash %}
+if [ -e /etc/SSHFLAG ]; then
+  /usr/sbin/update-rc.d -f ssh defaults
+  /bin/rm /etc/SSHFLAG
+  /sbin/shutdown -r now
+fi
+{% endhighlight %}
 
 kod parçası `exit0` komutunun hemen üstüne eklenir. Bu sayede Pi boot edildiğinde /etc dizininde SSHFLAG dosyası var mı diye bakar ve varsa ssh konfigurasyonlarını yapacak olan komutu çalıştırır. Bu komut ssh public key'lerini oluşturup ssh bağlantısının sağlanabileceği zemini hazırlar. Bundan sonra da SSHFLAG dosyasını dizinden siler ve sistemi yeniden başlatır. Daha sonraki bootlarda bu kısım çalışmaz.
 
@@ -85,6 +79,7 @@ Artık Pi'a SSH ile erişip login olunabilir, VNC kurularak uzak masaüstü kul
 
 SSH bağlantısı için:
 
-    
-    <code>ssh pi@192.168.1.67 --p</code>
-    password: raspberry
+{% highlight bash %}
+ssh pi@192.168.1.67 --p
+password: raspberry
+{% endhighlight %}
